[#41] FTDependencies 모듈 추가
