+++ Projects/App/Sources/App/AppDelegate.swift
@@ -12,6 +12,7 @@ import FirebaseMessaging
 import iamport_ios
 import KakaoSDKCommon
 
+import FTDependencies
 import FTStorageInterface
 import FTStorage
