+++ Projects/App/Sources/Feature/Login/SignUp/SignUpViewModel.swift
@@ -10,6 +10,7 @@ import Foundation
 import RxCocoa
 import RxSwift
 
+import FTDependencies
 import FTUtility
 
 final class SignUpViewModel: ViewModel {
