+++ Projects/App/Sources/Feature/Main/Chat/ChatViewModel.swift
@@ -10,6 +10,7 @@ import Foundation
 import RxCocoa
 import RxSwift
 
+import FTDependencies
 import FTStorageInterface
 import FTUtility
