+++ Projects/App/Sources/Feature/Main/ChatList/ChatRoomViewModel.swift
@@ -10,6 +10,7 @@ import Foundation
 import RxCocoa
 import RxSwift
 
+import FTDependencies
 import FTStorageInterface
 
 final class ChatRoomViewModel: ViewModel {
