+++ Projects/App/Sources/Feature/Main/Home/HomeViewModel.swift
@@ -10,6 +10,8 @@ import Foundation
 import RxCocoa
 import RxSwift
 
+import FTDependencies
+
 final class HomeViewModel: ViewModel {
     
     struct Input {
