+++ Projects/App/Sources/Feature/Main/TabBarViewController.swift
@@ -11,6 +11,8 @@ import RxCocoa
 import RxSwift
 import SnapKit
 
+import FTDependencies
+
 final class TabBarViewController: RxBaseViewController {
     
     // MARK: - UI Components
