+++ Projects/App/Sources/Feature/Main/WebView/PGWebViewModel.swift
@@ -10,6 +10,8 @@ import Foundation
 import RxCocoa
 import RxSwift
 
+import FTDependencies
+
 final class PGWebViewModel: ViewModel {
     struct Input {
         let validPayment: Observable<String>
